Library (wcs.h): Check PC and CD matrix when both are given
Until now, when the PC and CD matrix existed in the input FITS file, and
they conflicted (where written by non-Gnuastro software that store old
headers which conflict!) one of the two matrices would be used by WCSLIB
(probably the PC matrix), and in the end, we would only write the PC
matrix. But in some scenarios (particularly when the CDELT keywords were
removed by that software), the final PC and CD matrices would be different
and create unreasonable WCS-related operations.

With this commit, a check was added to see if both the CD and PC matrixes
exist or not. When they are both present, the final matrix produced by both
will be checked. When the two matrices are significantly different, a
complete warning is printed, describing the problem (and solution). But to
let the program continue (and because this problem is most probably caused
by software ignoring CDELT), it will convert the internal PC+CDELT matrices
to the values from the CD matrix.

This fixes bug #59459.

This bug was reported by Alberto Madrigal, and fixed with the help of
Sachin Kumar Singh.