Query: Shortcut for Gaia early Data release 3 now added
Until now, the most recent Gaia database was DR2. However, a few days ago
the early DR3 (eDR3) data has become public. The eDR3 could be accessed by
giving the full string to '--dataset'. But the short format is easier to
remember and type.

With this commit, a shortcut has been added for eDR3 ('--dataset=edr3'). I
also found out that we weren't actually checking if any sky coverage was
calculated or not! So a test has been added.