Library (kdtree): corrected visualization in book, refactored library
The visualized output of the description k-d tree part of the book didn't
correspond to the actual output k-d tree from 'gal_kdtree_create': while
the input coordinates were the same as those in Wikipedia, the root node in
our run would be different from Wikipedia. So when a user would run our
code, and get a different k-d tree they would be confused.

With this commit, the output of 'gal_kdtree_create' is now used in the
visualization at the start of the k-d tree library part of the book and a
footnote has been added to explain the small difference with the k-d tree
from Wikipedia. The examples with the two k-d tree functions have also
become much more realistic (with the k-d tree root written as a FITS
keyword and later read).

Also, the partition making subroutine within the k-d tree creation function
is now separated from 'kdtree_median_find' and made into a seperate
function ('kdtree_make_partition'). Comments are improves a bit too.