Segment: will not consume excessive RAM when there are many clumps
Until now, when the input detection map had many (on the scale of hundreds
of thousands) clumps within one detection, it would consume an unreasonable
amount of RAM (which could exceed even the nonvolatile memory on a system!
on the order of 1.5 Terrabytes in one of our examples!).

This happened when working on a wide field of view image near the disk of
the milky way (which has MANY stars), we noticed that one detection
included +600000 clumps and during its processing, Segment needed more than
1Tb of space which wasn't available in the RAM. It then tried to make a
memory-mapped file, but the storage wasn't sufficient to host that either
and Segment crashed.

The main problem was in the adjacency matrix step of Segment: until now if
there were N clumps in a detection, it would construct a NxN adjacency
matrix to find the connected clumps (which is why the memory usage went so
high).

With this commit, the new 'gal_binary_connected_adjacency_list' has been
added to avoid this problem. This new function will produce an identical
output to 'gal_binary_connected_adjacency_matrix', but it won't use a
matrix, it will use a list of indexs. It can be slightly slower than using
the adjacency matrix but consumes FAR LESS memory when large numbers are
involved.

This bug was found with the help of Samane Raji.