Book (tutorial): added box for changing locale
Until now when a system's locale didn't use the point ('.') as a decimal
separator, the resulting numbers wouldn't be interpretted properly and give
false results. In particular these two lines in the tutorial:

echo $n $r | awk '{print $1 * ($2^2) * 3600}'

and

for z in $(seq 0.1 0.1 5); do

With this commit, a box has been added in the first place such cases may
come up and a short description of locales is given with known issues and
how to solve this problem. Also, at the start of the code boxes in the
"Angular coverage on the sky" and "Cosmological coverage", a line was added
to set 'LANG=C' and avoid potential confusions.

This was reported by Andrés García-Serra Romero and Carlos Morales Socorro.