First implementation of k-d tree matching, not complete
The first library implementation of the k-d tree matching has been done:
the new 'gal_match_kdtree' function will find the nearest point in the
first catalog to each point in the second catalog in parallel. But it still
needs to do some post-processing to make an output in the same format as
'gal_match_coordinates'.

Some points on the previous commit that have been fixed here. Please pay
close attention to these, run with 'git log -p -1' to see my changes.

 - As we had defined in the task, 'match_catalog_kdtree_build' should only
   build the k-d tree, write it into a file and abort. So it doesn't need
   to return any dataset!

 - The case when the user doesn't specify an '--output' wasn't planned (a
   default name).

 - The comments I had originally put in 'match_catalog_kdtree' explained
   that the input coordinates are already read into memory (as
   'p->cols1'). So there was no need to re-read the input in
   'match_catalog_kdtree_build'!!!

 - Again, in the comments I had originally put in 'match_catalog_kdtree', I
   had explained that the value of 'p->kdtreemode' already contains the
   proper macro for the k-d tree mode (done in 'ui.c' at the very start of
   the program). There was no need to re-compare the strings in
   'match_catalog_kdtree'.