MakeCatalog: new --halfmaxradius column, obs suffix removed
The '--fwhmobs' and '--halfradiusobs' options were one of the first ones we
added in the '--halfXXX' options. But as we added more, the names became
more standard like '--halfsumarea' or '--halfsumsb', or '--halfmaxarea' and
etc. However, until now, '--halfradiusobs' still had a differently
formatted name which could be confusing. Also, both had an 'obs' suffix
which was redundant (because MakeCatalog doesn't do any parametric fitting
yet).

With this commit, '--halfradiusobs' is renamed to '--halfsumradius' to have
a similar format to the rest and avoid confusion and '--fwhmobs' is renamed
to '--fwhm'.

Also, for the sake of symmetry between the '--halfsum*' and '--halfmax*'
options, the new '--halfmaxradius' column has been added with this commit.