Library (dimension.h): new elliptical distance function
Until now, Gnuastro's MakeProfiles program would internally calculate the
elliptical distance of a point from the center of an ellipse. But it was
too deeply intertwined with its internal structure.

With this commit, a copy of that function has been brought into the
library, enabling other programs to easily use this feature also (as well
as users in their custom programs).