Crop: small limit of check for wcsmode crop decreased to 1
When Crop is in WCS mode, it will check to see if the given width is
reasonable. This is because it often happens that people confuse degrees
with arcseconds or vice-versa. There are two limits that cause Crop to
abort: a smaller limit and a larger limit.

Until now, the smaller width (after conversion to pixels) was 3 pixels. But
3 pixels is not unreasonable, 1 pixel is. With this commit, if the width is
smaller than 1 pixel an error is printed.