library (fits.h): new function to check if data is HEALPix
2D HEALPix datasets (which have spherical "pixel"s and are used for
whole-sky datasets like those of Plank) are stored as 1D arrays (mostly a
table column). Since there isn't any internal solution in Gnuastro yet to
convert 1D HEALPix datasets to a 2D image, their special storage would give
un-expected errors for the users that could be confusing. Also, WCSLIB
builds and installs a "utility" called 'HPXcvt' for this job, but most
people are unaware of it.

With this commit, a new function has been added to the Gnuastro library
called 'gal_fits_hdu_is_healpix'. This function is used in the core library
function to read FITS images and if the user's input is actually a 1D
HEALPix dataset, it will print an informative error and abort. In the error
message, it recomments WCSLIB's 'HPXcvt' program as a solution.

Also, in the FITS program, when no action is requested (and it prints the
basic information about every HDU), an extra comment column has been added
that will check for HEALPix. If the dataset has the standard HEALPix
keywords, the new comment column will mention this for the user to know.

This issue was found with the help of Ignacio Trujillo.