Crop: any WCS is acceptable for WCS-mode crops over a single image
Until now, the WCS limitations (for example being aligned with the
coordinate axises) that are necessary for cropping from many files also
existed when cropping from a single file which was very annoying (forcing
the user to use Table's column arithmetic to find the conversion, and then
doing an image-mode crop).

With this commit, when there is only a single image to crop from the given
coordinates, widths or polygons are converted to image coordinates at the
very start and Crop will do the rest of its processing as if it was in
image mode. Therefore allowing it to work with any WCS that is recognizable
by WCSLIB.