MakeCatalog: good sky subtraction on one-element-per-tile sky
Until now, when the given Sky image that should have been subtracted was a
single-element-per-tile tessellation (for example output of NoiseChisel's
'--oneelempertile' option), MakeCatalog would mistakenly subtract it from
the object label image! This was happening because MakeCatalog's original
tessellation is done over the objects image.

With this commit, the main macro-like function that does the parsing over
the tiles for Sky subtraction is also given the 'values' array and the Sky
value is subtracted from that (only using the objects image as a grid/guide
for the pixel positions).

This fixes bug #59765.