MakeCatalog: order-based measurements account for no clumps
Until now, the order-based measurements (like the median and etc) would
only write a NaN in the column if the full object wasn't present in the
values image. However, it can happen that the object is partially covered
while the clump(s) isn't. In such cases, MakeCatalog would crash (saying
that it can't allocate a zero-sized dataset).

With this commit, the necessary checks have been added to avoid this crash
and simply put a NaN in any of those clump columns.

In the process two other minor corrections were made:

 - The possibly printed message that says the Sky value has been subtracted
   from the values is now printed within the main set of messages. Until
   now, it was mistakenly printed even before the name of MakeCatalog!

 - The apostrophe was removed from the description of the minimum and
   maximum position columns because it would go as two apostrophes in the
   final table.