Statistics: 2D histograms can now be made as FITS images
Until now 2D histograms were built as tables, that relied on visualization
tools to plot. But the visualization step was slow and too
versatile. However by definition, a 2D histogram is actually a 2D image!

With this commit, the '--histogram2d' option of Statistics now takes an
argument: either '--histogram2d=table' (same old table format) or
'--histogram2d=image' (which is new). When an image 2D histogram is
requested, a WCS is also added to the FITS image, enabling easy viewing of
the plot. See the description in the newly added "2D histogram as an image"
section of the book for more.