Statistics: fixed 2D histogram segmentation fault
Until now when the input columns didn't have units, the 2D histogram
feature of Statistics would crash when creating the WCS.

With this commit a condition has been added to not assume units when
creating the WCS, and only use them when one is defined. Also, following
this check, I noticed several memory leakage situations throughout
statistics (and in the WCS library) that have now been fixed.

This bug was reported by Samane Raji.