Book: improved documentation of the match.h library
Until now, the documentation of the 'gal_match_coordinates' function didn't
really describe the format of the output too well and was very
long. However, we will soon have a new k-d tree based matching function and
the formats of its inputs and outputs are the same as this function (to
allow users to easy switch between the two).

With this commit, the description of the inputs and outputs of
'gal_match_coordinates' have been brought outside/before the function, and
the description of this function only contains the arguments related to its
particular matching algorithm.